+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Latest
+- Added weekly news, showing all the statistics for a given week.
 
 ## 2025.05.16
 - Changed the News page so it shows news for the past 7 days, not the current week.